Resolved CID 351558 (#1 of 1): Invalid type in argument to printf format specifier
invalid_type: Argument SpellMgr::getInstance()->getSpellInfoMap()->size() to format specifier %u was expected to have type unsigned int but has type unsigned long

+++ src/world/Spell/Customization/HackFixes.cpp
@@ -645,7 +645,7 @@ void SpellMgr::applyHackFixes()
 {
     //Updating spell.dbc
 
-    LogNotice("World : Processing %u spells...", sSpellMgr.getSpellInfoMap()->size());
+    LogNotice("World : Processing %u spells...", static_cast<uint32_t>(sSpellMgr.getSpellInfoMap()->size()));
 
     SpellInfo* sp = nullptr;
